Narrow GBA equipment validity checks
Apply the equipment ID's 16-bit semantic width at the validity branch, reproducing retail's explicit clrlwi before the unsigned zero comparison.\n\nConstraint: The getter retains its full-width ABI while the item ID remains 16-bit.\nRejected: Revert the getter declaration to u16 | regressed the preceding equipment visibility path\nConfidence: high\nScope-risk: narrow\nDirective: Keep ABI width and semantic item width separated at the call site.\nTested: ninja build/GC6E01/src/game/gbaCommunication_candidate_800965C8_gc20.o; ninja -j4 build/GC6E01/report.json; fn_800965C8 fuzzy 97.88221% -> 98.14664%\nNot-tested: full link/runtime

+++ src/game/gbaCommunication.c
@@ -2952,7 +2952,7 @@ void fn_800965C8(void* window, u8* sprite) {
         break;
     case 0x551:
         value = pokemonGetSoubiItemDataId(pokemon);
-        if (value != 0) {
+        if ((u16)value != 0) {
             msgctrlSetValue(0x2D, (u16)value);
             fn_800FBB34(0, 0, *(s16*)(sprite + 0x54), *(s16*)(sprite + 0x56), color, 0x2BD3);
         }
